# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /discover/people:
    get:
      operationId: discoverPeople
      tags: [intelligence]
      summary: Discover recently added people with source-backed facets
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
        - name: country
          in: query
          required: false
          schema: { type: string }
        - name: organization
          in: query
          required: false
          schema: { type: string }
        - name: industry
          in: query
          required: false
          schema: { type: string }
        - name: dataset
          in: query
          required: false
          schema: { type: string }
        - name: program
          in: query
          required: false
          schema: { type: string }
        - name: sort
          in: query
          required: false
          schema: { type: string, enum: [recent, name], default: recent }
        - name: limit
          in: query
          required: false
          schema: { type: number, minimum: 1, maximum: 100, default: 24 }
      responses:
        "200":
          description: Recently added people and available source facets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverPeopleResponse"
...
  /entities/{id}/coverage:
    get:
      operationId: getEntityCoverage
      tags: [intelligence]
      summary: Get unverified news context for an entity
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: refresh
          in: query
          required: false
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Related coverage and feed availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverageResponse"
  /batch-screen:
    post:
      operationId: batchScreen
      tags: [intelligence]
      summary: Screen a CSV or newline-separated list of names
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchScreenInput"
      responses:
        "200":
          description: Batch screening results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchScreenResponse"
  /watchlists:
    get:
      operationId: listWatchlists
      tags: [research]
      responses:
        "200":
          description: Watchlists
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Watchlist" }
    post:
      operationId: createWatchlist
      tags: [research]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateWatchlistInput" }
      responses:
        "201":
          description: Created watchlist
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Watchlist" }
  /watchlists/{id}/entities:
    post:
      operationId: addWatchlistEntity
      tags: [research]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WatchlistEntityInput" }
      responses:
        "201":
          description: Watched entity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WatchlistEntity" }
  /watchlists/{id}/events:
    get:
      operationId: getWatchlistEvents
      tags: [research]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Watch events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WatchEvent" }
  /watchlists/{id}/check:
    post:
      operationId: checkWatchlist
      tags: [research]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Watchlist check result
          content:
            application/json:
              schema:
                type: object
                required: [checked, changes]
                properties:
                  checked: { type: number }
                  changes: { type: number }
  /cases:
    get:
      operationId: listCases
      tags: [research]
      responses:
        "200":
          description: Investigation cases
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ResearchCase" }
    post:
      operationId: createCase
      tags: [research]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCaseInput" }
      responses:
        "201":
          description: Created case
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResearchCase" }
  /cases/{id}:
    get:
      operationId: getCase
      tags: [research]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Investigation case
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CaseDetail" }
  /cases/{id}/entities:
    post:
      operationId: addCaseEntity
      tags: [research]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CaseEntityInput" }
      responses:
        "201":
          description: Added case entity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CaseEntity" }
  /cases/{id}/notes:
    post:
      operationId: addCaseNote
      tags: [research]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CaseNoteInput" }
      responses:
        "201":
          description: Added case note
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CaseNote" }
  /entities/{id}/export:
    get:
      operationId: exportEntity
      tags: [research]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [json, csv, pdf] }
      responses:
        "200":
          description: Entity export
          content:
            application/json: {}
            text/csv: {}
            application/pdf: {}
  /entities/{id}/network/export:
    get:
      operationId: exportEntityNetwork
      tags: [research]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: depth
          in: query
          required: false
          schema: { type: number, minimum: 1, maximum: 2, default: 2 }
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [json, csv, pdf] }
      responses:
        "200":
          description: Network export
          content:
            application/json: {}
            text/csv: {}
            application/pdf: {}
  /batch-screen/export:
    post:
      operationId: exportBatchScreen
      tags: [research]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BatchExportInput" }
      responses:
        "200":
          description: Batch export
          content:
            application/json: {}
            text/csv: {}
            application/pdf: {}
  /entities/search:
    get:
      operationId: searchEntities
      tags: [intelligence]
      summary: Search sanctioned entities
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: limit
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 25
            default: 10
      responses:
        "200":
          description: Search matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EntitySearchResult"
  /entities/{id}:
    get:
      operationId: getEntity
      tags: [intelligence]
      summary: Get a sanctioned entity dossier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entity dossier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
        "404":
          description: Entity not found
  /entities/{id}/network:
    get:
      operationId: getEntityNetwork
      tags: [intelligence]
      summary: Get an expandable entity relationship network
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: depth
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 2
            default: 1
      responses:
        "200":
          description: Network nodes and edges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkGraph"
  /paths:
    get:
      operationId: findEntityPath
      tags: [intelligence]
      summary: Find the shortest relationship path between two entities
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shortest path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PathResult"
components:
  schemas:
...
    EntitySearchResult:
      type: object
      required: [id, name, schemaType, score, datasets, matchReasons]
      properties:
        id: { type: string }
        name: { type: string }
        schemaType: { type: string }
        score: { type: number }
        matchReasons:
          type: array
          items: { $ref: "#/components/schemas/MatchReason" }
        datasets:
          type: array
          items: { type: string }
        country:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
...
    DiscoverPerson:
      type: object
      required: [id, name, schemaType, aliases, datasets, countries, organizations, industries, programs, roles, addedAt]
      properties:
        id: { type: string }
        name: { type: string }
        schemaType: { type: string }
        aliases:
          type: array
          items: { type: string }
        datasets:
          type: array
          items: { type: string }
        countries:
          type: array
          items: { type: string }
        organizations:
          type: array
          items: { type: string }
        industries:
          type: array
          items: { type: string }
        programs:
          type: array
          items: { type: string }
        roles:
          type: array
          items: { type: string }
        birthDate: { type: ["string", "null"] }
        addedAt: { type: string }
        sourceUrl: { type: ["string", "null"] }
    DiscoverPeopleFacets:
      type: object
      required: [countries, organizations, industries, datasets, programs, roles]
      properties:
        countries:
          type: array
          items: { $ref: "#/components/schemas/DiscoverFacetOption" }
        organizations:
          type: array
          items: { $ref: "#/components/schemas/DiscoverFacetOption" }
        industries:
          type: array
          items: { $ref: "#/components/schemas/DiscoverFacetOption" }
        datasets:
          type: array
          items: { $ref: "#/components/schemas/DiscoverFacetOption" }
        programs:
          type: array
          items: { $ref: "#/components/schemas/DiscoverFacetOption" }
        roles:
          type: array
          items: { $ref: "#/components/schemas/DiscoverFacetOption" }
    DiscoverPeopleResponse:
      type: object
      required: [items, total, generatedAt, facets]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/DiscoverPerson" }
        total: { type: number }
        generatedAt: { type: string }
        facets: { $ref: "#/components/schemas/DiscoverPeopleFacets" }
    Entity:
      type: object
      required: [id, name, schemaType, aliases, datasets, properties, sources]
      properties:
        id: { type: string }
        name: { type: string }
        schemaType: { type: string }
        aliases:
          type: array
          items: { type: string }
        datasets:
          type: array
          items: { type: string }
        properties:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceCitation"
    SourceCitation:
      type: object
      required: [title, url]
      properties:
        title: { type: string }
        url: { type: string }
        publisher:
          type: ["string", "null"]
        category:
          type: string
          enum: [sanctions, sanctions_notice, court_proceeding, investigation, pep, adverse_media, corporate_registry, investigative_database]
        jurisdiction:
          type: ["string", "null"]
        sourceStatus:
          type: string
          enum: [official, normalized, journalistic, unverified_context, investigator_note]
        publishedAt:
          type: ["string", "null"]
        updatedAt:
          type: ["string", "null"]
        measureType:
          type: ["string", "null"]
        legalStatus:
          type: ["string", "null"]
...
    NetworkGraph:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/NetworkNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/NetworkEdge"
        sourceStatus:
          type: string
          enum: [available, setup_required, unavailable]
        message:
          type: ["string", "null"]
    PathResult:
      type: object
      required: [found, nodes, edges]
      properties:
        found: { type: boolean }
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/NetworkNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/NetworkEdge"
...
    MatchReason:
      type: object
      required: [label, detail]
      properties:
        label: { type: string }
        detail: { type: string }
    CoverageItem:
      type: object
      required: [id, feed, title, url, publisher, verification, fetchedAt]
      properties:
        id: { type: number }
        feed: { type: string }
        title: { type: string }
        url: { type: string }
        publisher: { type: string }
        summary: { type: ["string", "null"] }
        publishedAt: { type: ["string", "null"] }
        fetchedAt: { type: string }
        verification: { type: string }
    CoverageFeedStatus:
      type: object
      required: [feed, status]
      properties:
        feed: { type: string }
        status: { type: string }
        message: { type: ["string", "null"] }
    CoverageResponse:
      type: object
      required: [items, feeds]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CoverageItem" }
        feeds:
          type: array
          items: { $ref: "#/components/schemas/CoverageFeedStatus" }
    BatchScreenInput:
      type: object
      required: [csv]
      properties:
        csv: { type: string }
        limitPerName: { type: number, minimum: 1, maximum: 10, default: 3 }
...
    BatchScreenItem:
      type: object
      required: [input, status, matches]
      properties:
        input: { type: string }
        status: { type: string }
        matches:
          type: array
          items: { $ref: "#/components/schemas/BatchMatch" }
        message: { type: ["string", "null"] }
    BatchScreenResponse:
      type: object
      required: [items, generatedAt]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/BatchScreenItem" }
        generatedAt: { type: string }
    BatchExportInput:
      type: object
      required: [items, format]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/BatchScreenItem" }
        format: { type: string, enum: [json, csv, pdf] }
    Watchlist:
      type: object
      required: [id, name, createdAt, updatedAt, entities]
      properties:
        id: { type: number }
        name: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
        entities: { type: number }
    CreateWatchlistInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
    WatchlistEntityInput:
      type: object
      required: [entityId]
      properties:
        entityId: { type: string }
    WatchlistEntity:
      type: object
      required: [id, watchlistId, entityId, addedAt]
      properties:
        id: { type: number }
        watchlistId: { type: number }
        entityId: { type: string }
        addedAt: { type: string }
        lastCheckedAt: { type: ["string", "null"] }
    WatchEvent:
      type: object
      required: [id, watchlistEntityId, entityId, eventType, summary, detectedAt]
      properties:
        id: { type: number }
        watchlistEntityId: { type: number }
        entityId: { type: string }
        eventType: { type: string }
        summary: { type: string }
        before: { type: ["object", "null"], additionalProperties: true }
        after: { type: ["object", "null"], additionalProperties: true }
        detectedAt: { type: string }
        readAt: { type: ["string", "null"] }
    ResearchCase:
      type: object
      required: [id, name, description, createdAt, updatedAt, entityCount, noteCount]
      properties:
        id: { type: number }
        name: { type: string }
        description: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
        entityCount: { type: number }
        noteCount: { type: number }
    CreateCaseInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: ["string", "null"] }
    CaseEntityInput:
      type: object
      required: [entityId]
      properties:
        entityId: { type: string }
        note: { type: ["string", "null"] }
    CaseEntity:
      type: object
      required: [id, caseId, entityId, addedAt]
      properties:
        id: { type: number }
        caseId: { type: number }
        entityId: { type: string }
        note: { type: ["string", "null"] }
        addedAt: { type: string }
    CaseNoteInput:
      type: object
      required: [body]
      properties:
        body: { type: string }
    CaseNote:
      type: object
      required: [id, caseId, body, createdAt, updatedAt]
      properties:
        id: { type: number }
        caseId: { type: number }
        body: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    CaseDetail:
      allOf:
        - $ref: "#/components/schemas/ResearchCase"
        - type: object
          required: [entities, notes]
          properties:
            entities:
              type: array
              items: { $ref: "#/components/schemas/CaseEntity" }
            notes:
              type: array
              items: { $ref: "#/components/schemas/CaseNote" }